+++ noodles-bam/examples/bam_filter_name.rs
@@ -52,7 +52,7 @@ fn main() -> io::Result<()> {
     let header = reader.read_header()?;
 
     let stdout = io::stdout().lock();
-    let mut writer = bam::Writer::new(stdout);
+    let mut writer = bam::io::Writer::new(stdout);
 
     writer.write_header(&header)?;