+++ noodles-bam/src/io.rs
@@ -2,5 +2,6 @@
 
 pub mod indexed_reader;
 pub mod reader;
+pub mod writer;
 
-pub use self::{indexed_reader::IndexedReader, reader::Reader};
+pub use self::{indexed_reader::IndexedReader, reader::Reader, writer::Writer};